restoration pipeline: Fix FP16 performance regression in blend mask creation on Nvdia 50-series GPUs (#270)
**Description**
This PR addresses a FP16 performance regression observed during blend mask creation on v0.10.0, particularly on NVIDIA 50-series GPUs when processing large mosaic regions.

Previously, the blend mask was created using self.mosaic_restoration_model.dtype (FP16 in FP16 mode), followed by a .to(..., dtype=target_dtype) conversion.
Under FP16, this resulted in an expensive FP16 → FP32 conversion and device transfer, causing significantly increased CUDA utilization and major FPS drops (up to ~5× slower in some cases).

This change forces the blend mask to be created in FP32 directly:
```
clip_mask.to(device=self.device).float()
```
This avoids the costly intermediate conversion while preserving the existing behavior, since `target_dtype` is typically FP32 under default settings.

**Discussion:** https://codeberg.org/ladaapp/lada/issues/255

**Test:**
SFW videos
https://pixeldrain.com/u/6EkBHkyu
Test 1(steak_test1.mp4): Small-area mosaic
Test 2(steak_test2.mp4): Large-area mosaic

**Results:**
Testing with the Test 1 video shows little difference for small-area mosaic, but the difference is significant with Test 2 video.
The following test2 results are mostly from Windows builds, where the performance difference is clearly observable.
On RTX 20 / 30 / 40-series GPUs, it seems no significant performance difference was observed between FP16 and FP32.

On v0.10.0:
| GPU                | FP16 Time | FP32 Time | FP16 vs FP32 |
| ------------------ | --------- | --------- | ------------ |
| RTX 5050           | 7:40      | 1:39      | ~4.6× slower |
| RTX 5060           | 4:53      | 1:26      | ~3.4× slower |
| RTX 5070           | 3:03      | 1:26      | ~2.1× slower |
| RTX 5090  | 1:38      | 0:50      | ~2.0× slower |

Apply this change:
| GPU                | FP16 Time | FP32 Time |
| ------------------ | --------- | --------- |
| RTX 5050           | 1:28      | 1:39      |

Reviewed-on: https://codeberg.org/ladaapp/lada/pulls/270
Co-committed-by: FishYu <htps0763@gmail.com>

+++ lada/restorationpipeline/frame_restorer.py
@@ -196,7 +196,7 @@ class FrameRestorer:
             clip_mask = image_utils.unpad_image(clip_mask, pad_after_resize)
             clip_img = image_utils.resize(clip_img, orig_crop_shape[:2])
             clip_mask = image_utils.resize(clip_mask, orig_crop_shape[:2],interpolation=cv2.INTER_NEAREST)
-            blend_mask = mask_utils.create_blend_mask(clip_mask.to(device=self.device).to(dtype=self.mosaic_restoration_model.dtype)).to(device=clip_img.device, dtype=target_dtype)
+            blend_mask = mask_utils.create_blend_mask(clip_mask.to(device=self.device).float()).to(device=clip_img.device, dtype=target_dtype)
 
             blend(blend_mask, clip_img, orig_clip_box)
